requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
openai==1.3.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
openai==1.3.0
python-dotenv==1.0.0
alembic==1.13.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6